+++ pyvo/dal/tap.py
@@ -23,6 +23,7 @@ from ..io.vosi import tapregext as tr
 
 from ..utils.formatting import para_format_desc
 from ..utils.http import use_session
+from ..utils.prototype import prototype_feature
 import xml.etree.ElementTree
 import io
 
@@ -30,6 +31,14 @@ __all__ = [
     "search", "escape", "TAPService", "TAPQuery", "AsyncTAPJob", "TAPResults"]
 
 IVOA_DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
+
+# file formats supported by table upload and their corresponding MIME types
+TABLE_UPLOAD_FORMAT = {'tsv': 'text/tab-separated-values',
+                       'csv': 'text/csv',
+                       'FITSTable': 'application/fits'}
+# file formats supported by table create and their corresponding MIME types
+TABLE_DEF_FORMAT = {'VOSITable': 'text/xml',
+                    'VOTable': 'application/x-votable+xml'}
 
 
 def _from_ivoa_format(datetime_str):
@@ -101,7 +110,7 @@ class TAPService(DALService, AvailabilityMixin, CapabilityMixin):
 
     def __init__(self, baseurl, session=None):
         """
-        instantiate a Tablee Access Protocol service
+        instantiate a Table Access Protocol service
 
         Parameters
         ----------
@@ -115,7 +124,7 @@ class TAPService(DALService, AvailabilityMixin, CapabilityMixin):
         # Check if the session has an update_from_capabilities attribute.
         # This means that the session is aware of IVOA capabilities,
         # and can use this information in processing network requests.
-        # One such usecase for this is auth.
+        # One such use case for this is auth.
         if hasattr(self._session, 'update_from_capabilities'):
             self._session.update_from_capabilities(self.capabilities)
 
@@ -450,6 +459,127 @@ class TAPService(DALService, AvailabilityMixin, CapabilityMixin):
         for cap in capabilities:
             cap.describe()
             print()
+
+    @prototype_feature('cadc-tb-upload')
+    def create_table(self, name, definition, format='VOSITable'):
+        """
+        Creates a table in the catalog service.
+
+        Parameters
+        ----------
+        name: str
+            Name of the table in the TAP service
+        definition: stream (object with a read method)
+            Definition of the table
+        format: str
+            Format of the table definition (VOSITable or VOTable).
+        """
+        if not name or not definition:
+            raise ValueError(
+                'table name and definition required in create: {}/{}'.
+                format(name, definition))
+        if format not in TABLE_DEF_FORMAT.keys():
+            raise ValueError(
+                'Table definition file format {} not supported ({})'.
+                format(format, ' '.join(TABLE_DEF_FORMAT.keys())))
+
+        headers = {'Content-Type': TABLE_DEF_FORMAT[format]}
+        response = self._session.put('{}/tables/{}'.format(self.baseurl, name),
+                                     headers=headers,
+                                     data=definition)
+        response.raise_for_status()
+
+    @prototype_feature('cadc-tb-upload')
+    def remove_table(self, name):
+        """
+        Remove a table from the catalog service (Equivalent to drop command
+        in DB).
+
+        Parameters
+        ----------
+        name: str
+            Name of the table in the TAP service
+        """
+        if not name:
+            raise ValueError(
+                'table name required in : {}'.
+                format(name))
+
+        response = self._session.delete(
+            '{}/tables/{}'.format(self.baseurl, name))
+        response.raise_for_status()
+
+    @prototype_feature('cadc-tb-upload')
+    def load_table(self, name, source, format='tsv'):
+        """
+        Loads content to a table
+
+        Parameters
+        ----------
+        name: str
+            Name of the table
+        source: stream with a read method
+            Stream containing the data to be loaded
+        format: str
+            Format of the data source: tab-separated values(tsv),
+            comma-separated values (csv) or FITS table (FITSTable)
+        """
+        if not name or not source:
+            raise ValueError(
+                'table name and source required in upload: {}/{}'.
+                format(name, source))
+        if format not in TABLE_UPLOAD_FORMAT.keys():
+            raise ValueError(
+                'Table content file format {} not supported ({})'.
+                format(format, ' '.join(TABLE_UPLOAD_FORMAT.keys())))
+
+        headers = {'Content-Type': TABLE_UPLOAD_FORMAT[format]}
+        response = self._session.post(
+            '{}/load/{}'.format(self.baseurl, name),
+            headers=headers,
+            data=source)
+        response.raise_for_status()
+
+    @prototype_feature('cadc-tb-upload')
+    def create_index(self, table_name, column_name, unique=False):
+        """
+        Creates a table index in the catalog service.
+
+        Parameters
+        ----------
+        table_name: str
+            Name of the table
+        column_name: str
+            Name of the column in the table
+        unique: bool
+            True for unique index, False otherwise
+        """
+        if not table_name or not column_name:
+            raise ValueError(
+                'table and column names are required in index: {}/{}'.
+                format(table_name, column_name))
+
+        result = self._session.post('{}/table-update'.format(self.baseurl),
+                                    data={'table': table_name,
+                                          'index': column_name,
+                                          'unique': 'true' if unique
+                                          else 'false'},
+                                    allow_redirects=False)
+
+        if result.status_code == 303:
+            job_url = result.headers['Location']
+            if not job_url:
+                raise RuntimeError(
+                    'table update job location missing in response')
+            # run the job
+            job = AsyncTAPJob(job_url, session=self._session)
+            job = job.run().wait()
+            job.raise_if_error()
+            # TODO job.delete()
+        else:
+            raise RuntimeError(
+                'BUG: table update expected status 303 received {}'.
+                format(result.status_code))
 
 
 class AsyncTAPJob: